Kaidee/requirements.txt
Flask-Cors==3.0.2
Flask-Login==0.4.0
Flask-MySQL==1.4.0
Flask-Session==0.3.1
image==1.5.5
Jinja2==2.9.5
magickwand==0.2
mysql==0.0.1
MySQL-python==1.2.5
pytest==3.0.7
pytest-flask==0.10.0
python-dateutil==2.6.0
requests==2.13.0
responses==0.5.1
Wand==0.4.4
